% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mCSB.r
\name{cvCSB}
\alias{cvCSB}
\title{Coefficient of Variation of the Contracted Symmetric Beta Distribution}
\usage{
cvCSB(shape, q)
}
\arguments{
\item{shape}{Positive shape parameter \eqn{\alpha}.}

\item{q}{Positive contraction parameter.}
}
\value{
A numeric value.
}
\description{
Computes the theoretical coefficient of variation of the CSB distribution.
}
\examples{
cvCSB(shape = 2, q = 5)

}
