% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mCSB.r
\name{kurtCSB}
\alias{kurtCSB}
\title{Kurtosis of the Contracted Symmetric Beta Distribution}
\usage{
kurtCSB(shape, q, excess = TRUE)
}
\arguments{
\item{shape}{Positive shape parameter \eqn{\alpha}.}

\item{q}{Positive contraction parameter.}

\item{excess}{Logical. If \code{TRUE}, the excess kurtosis is returned;
otherwise, the ordinary kurtosis is returned.}
}
\value{
A numeric value.
}
\description{
Computes the theoretical kurtosis or excess kurtosis of the CSB distribution.
}
\examples{
kurtCSB(shape = 2, q = 5)
kurtCSB(shape = 2, q = 5, excess = FALSE)

}
