% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mCSB.r
\name{meanCSB}
\alias{meanCSB}
\title{Mean of the Contracted Symmetric Beta Distribution}
\usage{
meanCSB(shape, q)
}
\arguments{
\item{shape}{Positive shape parameter \eqn{\alpha}.}

\item{q}{Positive contraction parameter.}
}
\value{
A numeric value.
}
\description{
Computes the theoretical mean of the CSB distribution.
}
\examples{
meanCSB(shape = 2, q = 5)

}
