% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mCSB.r
\name{skewCSB}
\alias{skewCSB}
\title{Skewness of the Contracted Symmetric Beta Distribution}
\usage{
skewCSB(shape, q)
}
\arguments{
\item{shape}{Positive shape parameter \eqn{\alpha}.}

\item{q}{Positive contraction parameter.}
}
\value{
A numeric value.
}
\description{
Computes the theoretical skewness coefficient of the CSB distribution.
}
\examples{
skewCSB(shape = 2, q = 5)

}
