% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mCSB.r
\name{varCSB}
\alias{varCSB}
\title{Variance of the Contracted Symmetric Beta Distribution}
\usage{
varCSB(shape, q)
}
\arguments{
\item{shape}{Positive shape parameter \eqn{\alpha}.}

\item{q}{Positive contraction parameter.}
}
\value{
A numeric value.
}
\description{
Computes the theoretical variance of the CSB distribution.
}
\examples{
varCSB(shape = 2, q = 5)

}
